rename `src/inputs.ts` into `src/action.ts`
