+++ src/main.ts
@@ -1,11 +1,11 @@
 import * as core from "@actions/core";
 import * as exec from "@actions/exec";
+import * as action from "./action";
 import * as gcovr from "./gcovr";
-import { parseInputs } from "./inputs";
 
 async function run(): Promise<void> {
   try {
-    const inputs = parseInputs();
+    const inputs = action.parseInputs();
     core.startGroup("Install gcovr");
     await exec.exec("pip3 install gcovr");
     core.endGroup();
